Triangulation: simplify namespace
